requirements.txt
google-genai==1.12.1
requests==2.32.3
python-dotenv==1.0.1
